Separate post calls to a rest service, write tests for controller
